+++ packages/basehub/src/next/toolbar/client-conditional-renderer.tsx
@@ -25,7 +25,7 @@ export const ClientConditionalRenderer = ({
   }) => Promise<{ status: number; response: object }>;
   disableDraftMode: () => Promise<void>;
   revalidateTags: (o: { tags: string[] }) => Promise<{ success: boolean }>;
-  getLatestBranches: (o: { bshbPreviewToken: string }) => Promise<{
+  getLatestBranches: (o: { bshbPreviewToken: string | undefined }) => Promise<{
     status: number;
     response: LatestBranch[] | { error: string };
   }>;