+++ playground/CHANGELOG.md
@@ -1,5 +1,12 @@
 # playground
 
+## 0.0.164-canary.10
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@8.0.0-canary.10
+
 ## 0.0.164-canary.9
 
 ### Patch Changes